Nexi - Host/requirements.txt
discord.py
flask
gunicorn
psutil